refactored all imports and styles
