+++ .eslintrc.json
@@ -5,6 +5,7 @@
     "semi": 0,
     "no-alert": 0,
     "react/no-string-refs": 0,
-    "react/no-access-state-in-setstate": 0
+    "react/no-access-state-in-setstate": 0,
+    "react/no-unused-state": 0
   }
 }
