+++ src/components/CodeEditorCard.jsx
@@ -3,26 +3,30 @@ import CodeEditor from './CodeEditor'
 import PropTypes from 'prop-types'
 import {DragSource} from 'react-dnd'
 import db from '../firestore'
-
 import {withStyles} from '@material-ui/core/styles'
-import Card from '@material-ui/core/Card'
-import CardContent from '@material-ui/core/CardContent'
-import Typography from '@material-ui/core/Typography'
-import Dialog from '@material-ui/core/Dialog'
-import EditIcon from '@material-ui/icons/Edit'
-import DialogActions from '@material-ui/core/DialogActions'
-import DialogContent from '@material-ui/core/DialogContent'
-import Snackbar from '@material-ui/core/Snackbar'
-import DialogTitle from '@material-ui/core/DialogTitle'
-import DoneIcon from '@material-ui/icons/Done'
-import CancelIcon from '@material-ui/icons/Cancel'
-import Button from '@material-ui/core/Button'
-import Input from '@material-ui/core/Input'
-import FormControl from '@material-ui/core/FormControl'
-import FormHelperText from '@material-ui/core/FormHelperText'
-import Select from '@material-ui/core/Select'
-import Switch from '@material-ui/core/Switch'
-import Notification from './Notification'
+import styles from '../styles/CodeEditorCardStyles'
+import {
+  Card,
+  CardContent,
+  Typography,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  Snackbar,
+  DialogTitle,
+  Button,
+  Input,
+  FormControl,
+  FormHelperText,
+  Select,
+  Switch
+} from '@material-ui/core/'
+import {
+  Edit as EditIcon,
+  Done as DoneIcon,
+  Cancel as CancelIcon
+} from '@material-ui/icons/'
+import {Notification} from '../imports'
 
 const codeEditorSource = {
   beginDrag(props) {
@@ -37,29 +41,6 @@ function collect(connect, monitor) {
     isDragging: monitor.isDragging()
   }
 }
-
-const styles = theme => ({
-  card: {
-    minWidth: 275,
-    position: 'absolute'
-  },
-  label: {
-    textAlign: 'center'
-  },
-  button: {
-    margin: theme.spacing.unit,
-    marginLeft: '300px'
-  },
-  leftIcon: {
-    marginRight: theme.spacing.unit
-  },
-  rightIcon: {
-    marginLeft: theme.spacing.unit
-  },
-  iconSmall: {
-    fontSize: 15
-  }
-})
 
 class CodeEditorCard extends Component {
   constructor() {
