+++ src/components/Message.jsx
@@ -1,23 +1,7 @@
 import React, {Component} from 'react'
-
 import {withStyles} from '@material-ui/core/styles'
-import Typography from '@material-ui/core/Typography'
-
-const styles = theme => ({
-  message: {
-    padding: 7
-  },
-  sender: {
-    fontSize: 10
-  },
-  bubble: {
-    margin: 5,
-    padding: 5,
-    border: '1px solid ' + theme.palette.primary.main,
-    color: theme.palette.primary.main,
-    borderRadius: 5
-  }
-})
+import {Typography} from '@material-ui/core/'
+import styles from '../styles/MessageStyles'
 
 export class Message extends Component {
   render() {
