+++ src/components/Notification.jsx
@@ -1,16 +1,17 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
-import CheckCircleIcon from '@material-ui/icons/CheckCircle'
-import ErrorIcon from '@material-ui/icons/Error'
-import InfoIcon from '@material-ui/icons/Info'
-import CloseIcon from '@material-ui/icons/Close'
-import green from '@material-ui/core/colors/green'
-import amber from '@material-ui/core/colors/amber'
+import {
+  CheckCircle as CheckCircleIcon,
+  Error as ErrorIcon,
+  Info as InfoIcon,
+  Close as CloseIcon,
+  Warning as WarningIcon
+} from '@material-ui/icons/'
 import IconButton from '@material-ui/core/IconButton'
 import SnackbarContent from '@material-ui/core/SnackbarContent'
-import WarningIcon from '@material-ui/icons/Warning'
 import {withStyles} from '@material-ui/core/styles'
+import styles from '../styles/NotificationStyles'
 
 const variantIcon = {
   success: CheckCircleIcon,
@@ -18,34 +19,6 @@ const variantIcon = {
   error: ErrorIcon,
   info: InfoIcon
 }
-const styles = theme => ({
-  success: {
-    backgroundColor: green[600]
-  },
-  error: {
-    backgroundColor: theme.palette.error.dark
-  },
-  info: {
-    backgroundColor: theme.palette.primary.dark
-  },
-  warning: {
-    backgroundColor: amber[700]
-  },
-  icon: {
-    fontSize: 20
-  },
-  iconVariant: {
-    opacity: 0.9,
-    marginRight: theme.spacing.unit
-  },
-  message: {
-    display: 'flex',
-    alignItems: 'center'
-  },
-  margin: {
-    margin: theme.spacing.unit
-  }
-})
 
 function Notification(props) {
   const {classes, className, message, onClose, variant, ...other} = props
