+++ src/components/Output.jsx
@@ -1,22 +1,15 @@
 import React, {Component} from 'react'
 import PropTypes from 'prop-types'
 import {withStyles} from '@material-ui/core/styles'
-import ExpansionPanel from '@material-ui/core/ExpansionPanel'
-import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary'
-import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails'
-import Typography from '@material-ui/core/Typography'
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
+import {
+  ExpansionPanel,
+  ExpansionPanelDetails,
+  ExpansionPanelSummary,
+  Typography
+} from '@material-ui/core/'
+import {ExpandMore as ExpandMoreIcon} from '@material-ui/icons/'
 import limitEval from '../sanitize'
-
-const styles = theme => ({
-  root: {
-    width: '100%'
-  },
-  heading: {
-    fontSize: theme.typography.pxToRem(15),
-    fontWeight: theme.typography.fontWeightRegular
-  }
-})
+import styles from '../styles/OutputStyles'
 
 class Output extends Component {
   constructor(props) {
@@ -44,21 +37,21 @@ class Output extends Component {
     limitEval(
       input,
       (success, returnValue) => {
-        if (success) {
-          this.setState({
-            output: returnValue
-          })
-        } else {
-          try {
+        try {
+          if (success) {
+            this.setState({
+              output: returnValue
+            })
+          } else {
             this.setState({
               output:
                 'The code takes too long to run... Is there an infinite loop?'
             })
-          } catch (error) {
-            this.setState({
-              output: error.message
-            })
           }
+        } catch (error) {
+          this.setState({
+            output: error.message
+          })
         }
       },
       3000
