+++ src/components/VideoComponent.jsx
@@ -2,14 +2,8 @@ import React, {Component} from 'react'
 import Video from 'twilio-video'
 import axios from 'axios'
 import {withStyles} from '@material-ui/core/styles'
-import Switch from '@material-ui/core/Switch'
-
-const styles = () => ({
-  root: {
-    display: 'flex',
-    flexDirection: 'column'
-  }
-})
+import {Switch} from '@material-ui/core/'
+import styles from '../styles/VideoComponentStyles'
 
 export class VideoComponent extends Component {
   constructor(props) {
@@ -82,7 +76,6 @@ export class VideoComponent extends Component {
 
   roomJoined(room) {
     // Called when a participant joins a room
-    // console.log("Joined as '" + this.state.identity + "'")
     this.setState({
       activeRoom: room,
       localMediaAvailable: true,
@@ -98,20 +91,16 @@ export class VideoComponent extends Component {
 
     // Attach the Tracks of the Room's Participants.
     room.participants.forEach(participant => {
-      // console.log("Already in Room: '" + participant.identity + "'")
       var previewContainer = this.refs.remoteMedia
       this.attachParticipantTracks(participant, previewContainer)
     })
 
     // When a Participant joins the Room, log the event.
-    room.on('participantConnected', participant => {
-      // console.log("Joining: '" + participant.identity + "'")
-    })
+    room.on('participantConnected', participant => {})
 
     // When a Participant adds a Track, attach it to the DOM.
     //'trackAdded'
     room.on('trackSubscribed', (track, participant) => {
-      // console.log(participant.identity + ' added track: ' + track.kind)
       var previewContainer = this.refs.remoteMedia
       this.attachTracks([track], previewContainer)
       this.setState({otherpersoninRoom: true})
@@ -124,16 +113,13 @@ export class VideoComponent extends Component {
 
     // When a Participant removes a Track, detach it from the DOM.
     //'trackRemoved'
-    // room.on('trackUnsubscribed', (track, participant) => {
     room.on('trackRemoved', (track, participant) => {
-      // console.log(participant.identity + ' removed track: ' + track.kind)
       this.detachTracks([track])
       this.setState({otherpersoninRoom: false})
     })
 
     // When a Participant leaves the Room, detach its Tracks.
     room.on('participantDisconnected', participant => {
-      // console.log("Participant '" + participant.identity + "' left the room")
       this.detachParticipantTracks(participant)
     })
