+++ src/imports/aceThemes.js
@@ -1,0 +1,16 @@
+import 'brace/theme/monokai'
+import 'brace/theme/github'
+import 'brace/theme/tomorrow'
+import 'brace/theme/kuroir'
+import 'brace/theme/vibrant_ink'
+import 'brace/theme/xcode'
+import 'brace/theme/terminal'
+import 'brace/theme/twilight'
+import 'brace/theme/tomorrow_night_eighties'
+import 'brace/theme/mono_industrial'
+import 'brace/theme/eclipse'
+import 'brace/theme/chrome'
+import 'brace/theme/clouds_midnight'
+import 'brace/theme/merbivore_soft'
+import 'brace/theme/solarized_dark'
+import 'brace/theme/solarized_light'
