+++ src/imports/index.js
@@ -1,0 +1,47 @@
+import Canvas from '../components/Canvas'
+import Classroom from '../components/Classroom'
+import CodeEditor from '../components/CodeEditor'
+import CodeEditorCard from '../components/CodeEditorCard'
+import CreateRoom from '../components/CreateRoom'
+import HideBin from '../components/HideBin'
+import HomePage from '../components/HomePage'
+import Login from '../components/Login'
+import Message from '../components/Message'
+import Messaging from '../components/Messaging'
+import NavMenu from '../components/NavMenu'
+import Notepad from '../components/Notepad'
+import Notification from '../components/Notification'
+import Output from '../components/Output'
+import Profile from '../components/Profile'
+import RoomCard from '../components/RoomCard'
+import RoomContainer from '../components/RoomContainer'
+import RoomMembers from '../components/RoomMembers'
+import RoomStatusBar from '../components/RoomStatusBar'
+import Signup from '../components/Signup'
+import VideoCard from '../components/VideoCard'
+import VideoComponent from '../components/VideoComponent'
+
+export {
+  Canvas,
+  Classroom,
+  CodeEditor,
+  CodeEditorCard,
+  CreateRoom,
+  HideBin,
+  HomePage,
+  Login,
+  Message,
+  Messaging,
+  NavMenu,
+  Notepad,
+  Notification,
+  Output,
+  Profile,
+  RoomCard,
+  RoomContainer,
+  RoomMembers,
+  RoomStatusBar,
+  Signup,
+  VideoCard,
+  VideoComponent
+}
