+++ src/styles/CanvasStyle.js
@@ -1,0 +1,11 @@
+const styles = theme => ({
+  card: {
+    minWidth: 275,
+    position: 'absolute'
+  },
+  button: {
+    margin: theme.spacing.unit
+  }
+})
+
+export default styles
