+++ src/styles/ClassroomStyle.js
@@ -1,0 +1,25 @@
+const styles = theme => ({
+  root: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'space-between',
+    marginBottom: '264px'
+  },
+  room: {
+    flexGrow: 1,
+    height: '100vh'
+  },
+  paper: {
+    padding: theme.spacing.unit * 2,
+    textAlign: 'center',
+    color: theme.palette.text.secondary
+  },
+  card: {
+    minWidth: 275
+  },
+  pos: {
+    marginBottom: 12
+  }
+})
+
+export default styles
