+++ src/styles/CodeEditorCardStyles.js
@@ -1,0 +1,24 @@
+const styles = theme => ({
+  card: {
+    minWidth: 275,
+    position: 'absolute'
+  },
+  label: {
+    textAlign: 'center'
+  },
+  button: {
+    margin: theme.spacing.unit,
+    marginLeft: '300px'
+  },
+  leftIcon: {
+    marginRight: theme.spacing.unit
+  },
+  rightIcon: {
+    marginLeft: theme.spacing.unit
+  },
+  iconSmall: {
+    fontSize: 15
+  }
+})
+
+export default styles
