+++ src/styles/CreateRoomStyles.js
@@ -1,0 +1,7 @@
+const styles = theme => ({
+  margin: {
+    margin: theme.spacing.unit
+  }
+})
+
+export default styles
