+++ src/styles/HideBinStyles.js
@@ -1,0 +1,8 @@
+const styles = theme => ({
+  button: {
+    margin: theme.spacing.unit,
+    backgroundColor: 'none'
+  }
+})
+
+export default styles
