+++ src/styles/NavMenuStyles.js
@@ -1,0 +1,26 @@
+const styles = theme => ({
+  root: {
+    position: 'absolute'
+  },
+  flex: {
+    flexGrow: 1
+  },
+  menuButton: {
+    marginLeft: 0,
+    marginRight: 0
+  },
+  icon: {
+    margin: theme.spacing.unit * 2
+  },
+  list: {
+    width: 250
+  },
+  fullList: {
+    width: 'auto'
+  },
+  lowercase: {
+    textTransform: 'lowercase'
+  }
+})
+
+export default styles
