+++ src/styles/NotepadStyles.js
@@ -1,0 +1,14 @@
+const styles = () => ({
+  card: {
+    minWidth: 275,
+    position: 'absolute'
+  },
+  textArea: {
+    width: '100%',
+    maxHeight: '500px',
+    height: '-webkit-fill-available',
+    border: '1px dotted'
+  }
+})
+
+export default styles
