+++ src/styles/NotificationStyles.js
@@ -1,0 +1,32 @@
+import {green, amber} from '@material-ui/core/colors/'
+
+const styles = theme => ({
+  success: {
+    backgroundColor: green[600]
+  },
+  error: {
+    backgroundColor: theme.palette.error.dark
+  },
+  info: {
+    backgroundColor: theme.palette.primary.dark
+  },
+  warning: {
+    backgroundColor: amber[700]
+  },
+  icon: {
+    fontSize: 20
+  },
+  iconVariant: {
+    opacity: 0.9,
+    marginRight: theme.spacing.unit
+  },
+  message: {
+    display: 'flex',
+    alignItems: 'center'
+  },
+  margin: {
+    margin: theme.spacing.unit
+  }
+})
+
+export default styles
