+++ src/styles/OutputStyles.js
@@ -1,0 +1,11 @@
+const styles = theme => ({
+  root: {
+    width: '100%'
+  },
+  heading: {
+    fontSize: theme.typography.pxToRem(15),
+    fontWeight: theme.typography.fontWeightRegular
+  }
+})
+
+export default styles
