+++ src/styles/ParallaxStyle.js
