+++ src/styles/ProfileStyles.js
@@ -1,0 +1,32 @@
+import parallaxStyle from './ParallaxStyle'
+
+const styles = theme => ({
+  row: {
+    display: 'flex'
+  },
+  avatar: {
+    margin: 50
+  },
+  bigAvatar: {
+    width: 260,
+    height: 260
+  },
+  card: {
+    maxWidth: 345
+  },
+  margin: {
+    margin: theme.spacing.unit
+  },
+  textField: {
+    marginLeft: theme.spacing.unit,
+    marginRight: theme.spacing.unit,
+    width: 200
+  },
+  media: {
+    height: 0,
+    paddingTop: '56.25%' // 16:9
+  },
+  ...parallaxStyle
+})
+
+export default styles
