+++ src/styles/RoomCardStyles.js
@@ -1,0 +1,23 @@
+const styles = theme => ({
+  card: {
+    maxWidth: 300,
+    margin: '2%'
+  },
+  media: {
+    height: 0,
+    paddingTop: '56.25%' // 16:9
+  },
+  textField: {
+    marginLeft: theme.spacing.unit,
+    marginRight: theme.spacing.unit,
+    width: 200
+  },
+  rightIcon: {
+    marginLeft: theme.spacing.unit
+  },
+  button: {
+    margin: theme.spacing.unit
+  }
+})
+
+export default styles
