+++ src/styles/RoomContainerStyles.js
@@ -1,0 +1,19 @@
+const styles = () => ({
+  root: {
+    display: 'flex',
+    flexDirection: 'column',
+    marginBottom: 64
+  },
+  cardsHeading: {
+    display: 'flex',
+    justifyContent: 'space-evenly',
+    alignItems: 'center'
+  },
+  allCards: {
+    display: 'flex',
+    justifyContent: 'center',
+    flexWrap: 'wrap'
+  }
+})
+
+export default styles
