+++ src/styles/RoomMembersStyles.js
@@ -1,0 +1,10 @@
+const styles = () => ({
+  root: {
+    margin: '0 10px'
+  },
+  text: {
+    fontSize: 14
+  }
+})
+
+export default styles
