+++ src/styles/VideoCardStyles.js
@@ -1,0 +1,16 @@
+const styles = theme => ({
+  card: {
+    width: 275,
+    position: 'absolute'
+  },
+  margin: {
+    margin: theme.spacing.unit
+  },
+  textField: {
+    marginLeft: theme.spacing.unit,
+    marginRight: theme.spacing.unit,
+    width: 200
+  }
+})
+
+export default styles
