+++ src/styles/VideoComponentStyles.js
@@ -1,0 +1,8 @@
+const styles = () => ({
+  root: {
+    display: 'flex',
+    flexDirection: 'column'
+  }
+})
+
+export default styles
